Merge origin/master
Nathan - Modified nearly all View classes so the model layer information is correctly passed around.  This way we are not un-intentionally creating new model data.  Storms are populating.  Added a printMap function to the MapControl in order to view map and fulfill assignment requirements of double array and sum of a list.